#Libraries... Check description
So, if we want any other dirs other than apps and peripherals, we need
to add it to the platformio.ini file. Currently, this implementation
creates a library out of each folder within applications or peripherals.
I don't know if this will be our final solution, but it works well.
Platformio adds the compiled library files to the root directory of the
environment which means to access blinky.hpp you must do import
<blinky.hpp> instead of the quotes.